Minor: Move test with code
